Breaking Down Component Systems
Started  moving component logic to systems, so that components act as purely data containers, they must be paired with a system to peform operations on.

Created for loop initializer & deinitializer for engine subsytems.

Added context menu to scene objects panel. Delete button marks object for rubbish colletcion. Which deletes it before render submission. This is currently not entirey safe.